+++ fixtures/ext-types/external-crate/src/lib.rs
@@ -24,3 +24,7 @@ impl ExternalCrateInterface {
         self.sval.clone()
     }
 }
+
+pub enum ExternalCrateEnumInterface {
+    Whatever,
+}