+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
 ### What's Fixed
 - Kotlin: Fixed messages for error classes that inherit `Throwable`, but not `Exception`.
+- Fix UDL remote enums, now allowing `[Enum, Remote] interface { ... }`
+  (via [#2823](https://github.com/mozilla/uniffi-rs/issues/2823)).
 
 [All changes in [[UnreleasedUniFFIVersion]]](https://github.com/mozilla/uniffi-rs/compare/v0.32.0...HEAD).